Add : Mission 4 - 댓글 컴포넌트화, props로 데이터 전달 구현, 댓글에 좋아요 기능 구현.
