+++ src/components/Comment/Comment.js
@@ -1,8 +1,30 @@
-import React from 'react';
+import React, { useState } from 'react';
 import './Comment.scss';
 
-function Comment() {
-  return <div />;
+function Comment({ key, userName, commentText }) {
+  const [isLiked, setIsLiked] = useState(false); // 댓글에 isLiked 관련 state 생성.
+
+  const toggleIsLiked = () => {
+    setIsLiked(!isLiked); // isLiked에 들어간 booelan 값을 계속해서 toggle
+  };
+  return (
+    <div className="replyChunk">
+      <div className="replyItSelf">
+        <span className="shownReplyId">{userName}</span>
+        <span className="shownReplyText">{commentText}</span>
+      </div>
+      <img
+        className="imgHeartIcon"
+        src={
+          !isLiked
+            ? '/images/eojineOh/heart.png'
+            : '/images/eojineOh/redheart.png'
+        }
+        alt="Hearticon"
+        onClick={toggleIsLiked}
+      />
+    </div>
+  );
 }
 
 export default Comment;
