+++ src/styles/variables.scss
@@ -19,6 +19,11 @@
   justify-content: space-between;
 }
 
+@mixin widthHeight_100 {
+  width: 100%;
+  height: 100%;
+}
+
 @mixin flexStart {
   display: flex;
   justify-content: flex-start;
@@ -33,10 +38,7 @@
   height: 24px;
 }
 
-@mixin widthHeight_100 {
-  width: 100%;
-  height: 100%;
-}
+
 
 @mixin font_strong {
   font-weight: 600;
